pyproject.toml
[tool.poetry]
name = "back-sys"
version = "2.0.0"
description = "backend medicalapp"
authors = ["chaknogf <chaknogf@gmail.com>"]
license = "gpl"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.7.0"
anyio = "4.11.0"
bcrypt = "5.0.0"
certifi = "2025.11.12"
cffi = "2.0.0"
click = "8.3.1"
contourpy = "1.3.1"
cryptography = "46.0.3"
cycler = "0.12.1"
dnspython = "2.8.0"
ecdsa = "0.19.1"
email-validator = "2.3.0"
et-xmlfile = "2.0.0"
fastapi = "0.122.0"
fonttools = "4.56.0"
greenlet = "3.2.1"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
idna = "3.11"
kiwisolver = "1.4.8"
matplotlib = "3.10.7"
narwhals = "1.30.0"
numpy = "2.2.3"
openpyxl = "3.1.5"
packaging = "24.2"
pandas = "2.2.3"
passlib = "1.7.4"
pillow = "11.1.0"
plotly = "6.0.0"
psycopg2-binary = "2.9.11"
pyasn1 = "0.6.1"
pycparser = "2.23"
pydantic = "2.12.5"
pydantic-core = "2.41.5"
pyjwt = "2.10.1"
pymysql = "1.1.1"
pyparsing = "3.2.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.2.1"
python-jose = "3.5.0"
python-multipart = "0.0.20"
pytz = "2025.1"
rsa = "4.9.1"
six = "1.17.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.44"
starlette = "0.50.0"
typing-extensions = "4.15.0"
tzdata = "2025.1"
uvicorn = "0.38.0"
typing-inspection = "0.4.2"
aiosmtplib = "4.0.2"
annotated-doc = "0.0.4"
argon2-cffi = "25.1.0"
argon2-cffi-bindings = "25.1.0"
blinker = "1.9.0"
fastapi-mail = "1.5.8"
jinja2 = "3.1.6"
markupsafe = "3.0.3"
pydantic-settings = "2.12.0"
regex = "2025.11.3"
urllib3 = "2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
